third_party/WebKit/LayoutTests/fast/events/platform-wheelevent-paging-x-in-scrolling-div-expected.txt

PASS event.wheelDeltaY is 0
PASS event.wheelDeltaX is window.givenScrollLeft * -3
PASS event.wheelDelta is window.givenScrollLeft * -3
PASS div.scrollTop == window.expectedScrollTop && div.scrollLeft == window.expectedScrollLeft became true
PASS successfullyParsed is true

TEST COMPLETE
